UPDATING BOOK ITEMS FROM MyUploads : DONE :100: :fire:
Also did some changes in other files.
Next we need to fix the designs of XML files and need to validate the java files where inputs are taken from user.
Also need to add separate menu files for separate screens.
All the major logical parts are done in this application.

BOOKLOOP IS READY TO GO :fire: